feat(settings): extract the three General pages
Why: Appearance, Keyboard & actions and Updates move behind the registry.

Updates carries the only conditional rows in the set: on a Microsoft
Store install there is no check and no channel, so both are declared
when: "updatable" — the index has to gate on the same predicate the card
does, or a search for "update" would name a check that policy 10.2.5
forbids naming.

The guard test needed two corrections beyond the brief: declaredRowIds
no longer filters when-gated rows (the default render mounts Updates
under an unmanaged capability, so the gated rows genuinely appear —
excluding them made declared a strict subset of rendered), with a
dedicated Store-managed case covering the negative; and a new
label-equality assertion enforces the SettingRowMeta.label docstring's
existing claim, which nothing previously checked. The Appearance card
is marked dynamic: true because appearance.light/appearance.dark and
appearance.theme are a real, mutually-exclusive either/or in the
component's own JSX (follow-system vs fixed) — no single render can
show all eight declared rows at once, so the guard test's per-row DOM
check is skipped for this one card (registerCardRows still registers
all eight ids, so search-driven card visibility is unaffected); the
toggle itself stays covered by Settings.appearance.test.tsx's real
assertions.
