+++ src/features/settings/nav/pages.ts
@@ -1,5 +1,8 @@
 import { registerCardRows } from "@/features/settings/layout/SettingsCard";
+import * as appearance from "@/features/settings/pages/appearance";
 import * as diff from "@/features/settings/pages/diff";
+import * as keyboard from "@/features/settings/pages/keyboard";
+import * as updates from "@/features/settings/pages/updates";
 import type {
   SettingsGroupId,
   SettingsPageId,
@@ -26,12 +29,15 @@ export const GROUPS: readonly SettingsGroup[] = [
 ];
 
 export const PAGES: Record<SettingsPageId, SettingsPageModule> = {
+  "general.appearance": { meta: appearance.meta, Page: appearance.AppearancePage },
+  "general.keyboard": { meta: keyboard.meta, Page: keyboard.KeyboardPage },
+  "general.updates": { meta: updates.meta, Page: updates.UpdatesPage },
   "git.diff": { meta: diff.meta, Page: diff.DiffPage },
-  // Tasks 4-6 add the other nine. Until then the mapped type below fails to
+  // Task 6 adds the other six. Until then the mapped type below fails to
   // compile, which is the point: a missing page cannot be forgotten.
   //
   // The double cast (via `unknown`) is needed because TS's excess-property
-  // check refuses a direct cast from an object with 1 of 10 keys straight to
+  // check refuses a direct cast from an object with 4 of 10 keys straight to
   // `Record<SettingsPageId, …>` — "insufficient overlap". Scaffolding only:
   // Task 6 replaces this whole expression with a plain typed literal once all
   // ten keys are present, at which point the cast is not just unneeded but
