+++ src/screens/Settings.appearance.test.tsx
@@ -4,28 +4,10 @@
 import { fireEvent, render, screen, within } from "@testing-library/react";
 import { beforeEach, describe, expect, it } from "vitest";
 
-import { mockInvoke } from "@/test/invokeMock";
 import { WithDialogs, resetDialogs } from "@/test/dialog";
 import { pgPickOption, pgSelectTrigger, pgSelectValues } from "@/test/select";
 import { useSettingsStore } from "@/features/settings/useSettingsStore";
-import { SettingsScreen } from "./Settings";
-
-function mockRestOfSettings() {
-  mockInvoke("cli_shim_status", () => ({
-    installed: true,
-    shimPath: "/usr/local/bin/pgit",
-    target: "/usr/bin/platypusgit",
-    source: "package",
-    pathState: "onPath",
-  }));
-  mockInvoke("diagnostics_report", () => ({
-    logPath: "/tmp/platypusgit.log",
-    logExists: false,
-    logSizeBytes: 0,
-    environment: "host os=macos arch=aarch64 git=2.43.0",
-    version: "0.1.0",
-  }));
-}
+import { AppearancePage } from "@/features/settings/pages/appearance";
 
 /**
  * The Row whose label is `label`, so a control can be found by what it is FOR
@@ -48,7 +30,6 @@ function row(label: string): HTMLElement {
 beforeEach(() => {
   localStorage.clear();
   resetDialogs();
-  mockRestOfSettings();
   useSettingsStore.getState().reset();
   useSettingsStore.getState().syncSystemAppearance("dark");
 });
@@ -56,7 +37,7 @@ beforeEach(() => {
 function renderSettings() {
   render(
     <WithDialogs>
-      <SettingsScreen />
+      <AppearancePage />
     </WithDialogs>,
   );
 }
