+++ src/screens/settings.index.test.tsx
@@ -12,6 +12,8 @@ import { beforeEach, describe, expect, it } from "vitest";
 import { WithDialogs, resetDialogs } from "@/test/dialog";
 import { GROUPS, PAGES, PAGE_ORDER, resolvePageId, FIRST_PAGE } from "@/features/settings/nav/pages";
 import type { SettingsPageId } from "@/features/settings/nav/types";
+import { UpdatesPage } from "@/features/settings/pages/updates";
+import { useUpdateStore } from "@/features/update/useUpdateStore";
 
 /**
  * Pages actually in the registry.
@@ -28,12 +30,22 @@ function renderedRowIds(container: HTMLElement): string[] {
     .sort();
 }
 
-/** Row ids a page declares, minus gated ones (they are asserted separately). */
+/**
+ * Row ids a page declares.
+ *
+ * Gated rows (`when: "updatable"`) are included on purpose: the guard test
+ * mounts every page under the DEFAULT update capability (`capability: null`,
+ * `useUpdateStore`'s initial state), and `updatesManagedExternally(null)` is
+ * `false` — so `UpdatesPage`'s ordinary branch renders, and its gated rows
+ * genuinely appear in the DOM. Excluding them here would make `declared` a
+ * subset of `rendered` and hide a real drift. The Store-managed case (where
+ * the gated rows really are absent) gets its own case below, under "the
+ * update gate".
+ */
 function declaredRowIds(pageId: SettingsPageId): string[] {
   return PAGES[pageId].meta.cards
     .filter((c) => !c.dynamic)
     .flatMap((c) => c.rows)
-    .filter((r) => !r.when)
     .map((r) => r.id)
     .sort();
 }
@@ -41,6 +53,9 @@ function declaredRowIds(pageId: SettingsPageId): string[] {
 describe("settings registry", () => {
   beforeEach(() => {
     resetDialogs();
+    // Every case starts from the default, unmanaged capability — a case that
+    // sets it to "store-managed" (below) must not leak into its neighbours.
+    useUpdateStore.setState({ capability: null });
   });
 
   it("declares exactly the rows each page renders", () => {
@@ -55,6 +70,38 @@ describe("settings registry", () => {
         (id) => !isDynamicRow(pageId, id),
       );
       expect(rendered, `${pageId}: rendered rows`).toEqual(declaredRowIds(pageId));
+      unmount();
+    }
+  });
+
+  it("renders each row's label exactly as its meta declares", () => {
+    // Search matches on `meta.label` (SettingRowMeta's docstring promises this
+    // is enforced) — a drifted label would make a search result show text the
+    // user never actually sees on screen.
+    for (const pageId of REGISTERED) {
+      const { Page, meta } = PAGES[pageId];
+      const { container, unmount } = render(
+        <WithDialogs>
+          <Page />
+        </WithDialogs>,
+      );
+      for (const card of meta.cards) {
+        if (card.dynamic) continue;
+        for (const rowMeta of card.rows) {
+          const row = container.querySelector(
+            `[data-setting-id="${rowMeta.id}"]`,
+          );
+          // Absent under the default render (a gated row, or the other half
+          // of a follow-system/fixed toggle) — nothing to compare here; the
+          // both-directions id check above already pins presence.
+          if (!row) continue;
+          const labelText = row.firstElementChild?.firstElementChild
+            ?.textContent?.trim();
+          expect(labelText, `${pageId}/${rowMeta.id} label`).toBe(
+            rowMeta.label,
+          );
+        }
+      }
       unmount();
     }
   });
@@ -103,6 +150,38 @@ describe("settings registry", () => {
     expect(resolvePageId(undefined)).toBe(FIRST_PAGE);
     expect(resolvePageId(42)).toBe(FIRST_PAGE);
     expect(FIRST_PAGE).toBe("general.appearance");
+  });
+});
+
+describe("the update gate", () => {
+  beforeEach(() => {
+    resetDialogs();
+    useUpdateStore.setState({ capability: null });
+  });
+
+  it("declares the check and channel rows as gated", () => {
+    const rows = PAGES["general.updates"].meta.cards.flatMap((c) => c.rows);
+    expect(rows.find((r) => r.id === "updates.check")?.when).toBe("updatable");
+    expect(rows.find((r) => r.id === "updates.channel")?.when).toBe("updatable");
+    // The version row is NOT gated — it is what a Store install still shows.
+    expect(rows.find((r) => r.id === "updates.version")?.when).toBeUndefined();
+  });
+
+  it("drops the update check and channel on a Store install", () => {
+    // Store policy 10.2.5 makes NAMING an update check the violation, not just
+    // offering one — v0.4.0 failed certification on exactly this. The page must
+    // render neither row when updates are managed externally.
+    useUpdateStore.setState({ capability: "store-managed" });
+    const { container, unmount } = render(
+      <WithDialogs>
+        <UpdatesPage />
+      </WithDialogs>,
+    );
+    const ids = renderedRowIds(container);
+    expect(ids).toContain("updates.version");
+    expect(ids).not.toContain("updates.check");
+    expect(ids).not.toContain("updates.channel");
+    unmount();
   });
 });
